Lab2: change DetailCount to be long instead of int.

+++ lab2/start/MyCircle/MyCircle/IAsyncMessageRepository.cs
@@ -8,7 +8,7 @@ namespace MyCircle
     {
         Task AddAsync(CircleMessage message);
         Task<IEnumerable<CircleMessage>> GetRootsAsync(int count);
-        Task<int> GetDetailCountAsync(string id);
+        Task<long> GetDetailCountAsync(string id);
         Task<IEnumerable<CircleMessage>> GetDetailsAsync(string id);
     }
 }
